feat: pss-ctr-rec-root-contract-007 - Seal root-contract invariants for IMP-REC unlock
